Stim Manager - initial implementation of npoisson stimulus.
...  To facilitate this, give the StimManager a reference to the target manager because it will need the freedom to call special functions for getting apical and basal pointlists.  Minor adjustment to linear stim to be consistent then.  Eliminate local valueOf function in preference for the Map's version.  
TargetManager - introduce convenience functions for getting specific sections using a cell target.   

Target - patches for using Section targets and getting complete compartment pointlist (to include nil sections)

LoadBalance - takes control of the dummy nil section.  This is easier for a Target object to access.

Cell - keep a separate vector for the number of segments in each section so that after multisplit, the target manager can still generate the correct number of nils in the pointlist

HDF5reader & SynManager - included function to activate redirecting of stdout & stderr.  This is useful for debugging, but should be migrated into a separate mod file since it's current location is arbitrary and was taken just out of immediate convenience.

Patch Note: on bbplinsrv, calling pc.multisplit is causing a segmentation violation.  However, if it is called in a separate function, it works fine.  The bluegene doesn't have a problem with calling the function from the initial location, though.  Maybe contact Michael for guidance?

JGK



git-svn-id: https://bbpteam.epfl.ch/svn/bluebrain/branches/bglib1.5@593 ae2f4726-ae01-0410-bc15-f70e80f9baa8
--
NEURODAMUS-CORE Repo SHA: hpc/sim/neurodamus-core@484b55527610d063acf9f77b255d538090d11860