Minor modifications where biophys and geom_nseg should now be called
... automatically for typical usage.
That is, geom_nsec is called at the end of function setCellProperties where the resistance is modified.  Biophys is called in the init function when a morphology path is supplied.

Also, added TemplateCCell.hoc under bluebrains version control.  ModelManagement was originally given this file, but it should instead use svn export to download it.  By keeping the template in this repository, it will be easier to maintain/test as the bglib code changes.

JGK



git-svn-id: https://bbpteam.epfl.ch/svn/bluebrain/branches/bglib1.5@595 ae2f4726-ae01-0410-bc15-f70e80f9baa8
--
NEURODAMUS-CORE Repo SHA: hpc/sim/neurodamus-core@68773de17135244524f1195185458a314eeca588