update bglib1.5 to support Probabilistic Synapses with deterministic seeding.
This check in includes pending versions of the mod files (subject to change when officially committed)

Changes to the hoc files include giving the Cell a list object to hold Random objects which the synapses will reference (the Cell must hold a reference in hoc, or it will be dealloc'd)
Because of the increased need for access to the Cell object during synapse creation, the sequence of creation is modified.  Previously, the synapses were made and the objects held in a list until they could be passed to the cell object.  Now the synapse parameters are stored instead, so that the synapses are only created when the cell is available during the finalize stage.

JGK



git-svn-id: https://bbpteam.epfl.ch/svn/bluebrain/branches/bglib1.5@614 ae2f4726-ae01-0410-bc15-f70e80f9baa8
--
NEURODAMUS-CORE Repo SHA: hpc/sim/neurodamus-core@4b5f00a6f9a7a38b9a91df8382ca1f4fdfdad8ea