Minor changes to improve fault tolerance.
ElectrodeManager is now an optional arg to the StimulusManager.  This was causing problems with the pfe since it did not receive an update.  It's better that the pfe not need to pass a nil to the StimulusManager constructor in order to minimize what it needs to have knowledge about.  As the ElectrodeManager is now optional, an additional check is made for its existence before use, with an execerror issued if it is non-existent.

For the CellDistributor, when it is looking for corrected template names (where sp characters like [.-()] are converted to _), it was going only to the second line of the template file.  In recent development work, that was not valid and caused an error.  Now it searchs until it reaches the bgintemplate line in order to retrieve the NEURON valid name.

For synapses, Node.hoc will skip majority of synapse creation code if the config file has exactly 1 Connection block with a weight of 0.  This is useful for instances where no nrn.h5 file exists (such as model management).

JGK



git-svn-id: https://bbpteam.epfl.ch/svn/bluebrain/branches/bglib1.5@704 ae2f4726-ae01-0410-bc15-f70e80f9baa8
--
NEURODAMUS-CORE Repo SHA: hpc/sim/neurodamus-core@883b19cefa85c9e5a6c64e82b3c9eb374ee09a7a