Support added to compute LoadBalancing for different number of cpus.
...  This is restoring the old ability that had been available with the original bglib as requested.  It is handled via the BlueConfig by adding a new ProspectiveHosts option
Run default
{
               ... ...
     CircuitTarget MiniColumn_0
           RunMode LoadBalance
  ProspectiveHosts 32
}

An additional file is written, cxinfo_%d.txt which writes out the circuit target and the circuit path (nrnpath) in order to see if the load balancing info for the current run setup is still valid.  Example: I run with some circuit, MiniColumn_0.  LoadBalancing is computed the first time, and reused on subsequent runs.  If I change to MiniColumn_1, the cxinfo will differ, so LoadBalancing info is recomputed.  Note that if the contents of a Target change, this is not detected.

JGK



git-svn-id: https://bbpteam.epfl.ch/svn/bluebrain/branches/bglib1.5@811 ae2f4726-ae01-0410-bc15-f70e80f9baa8
--
NEURODAMUS-CORE Repo SHA: hpc/sim/neurodamus-core@7d059e3249a5d9366d2fd21687f06c36fbaddc10