Regarding JIRA ticket BBPBGLIB-150
For summation reports, we need a special handling for SEClamp and IClamp in order to be consistent with the sign of the elecrtical current.  As such, a scalar of -1 is applied to the reported var.

JGK
--
NEURODAMUS-CORE Repo SHA: hpc/sim/neurodamus-core@ebdc39272d46b653e64d1412d79bb91f38c49afc