Bug fix for TStim noise stimulation where amp continued after del+dur.
The reason is because the rand populated the stim vector, but we forgot to reset the amp
value back to 0 at the end.  Therefore, the IClamp continues with the last value that the
random stream assigned into the amp var.  This fix ensures that the last vector value is 0

JGK
--
NEURODAMUS-CORE Repo SHA: hpc/sim/neurodamus-core@05ee030ae62aa8ca3c53f9c79f80a07f98bf4a40