Update to Reporting to allow compartment surface area to be reported
since area is a function call, it is not addressable the same way as membrane voltage or currents.
As a workaround, the ALU object has a new 'constant' operation added.  This allows us to store the result
of the area into the scalar field, and then report on only the scalar value.

Perhaps a more elegant solution can be done in the future.

JGK

Change-Id: I6ba3bf132e988080fc8bdcac06b2a67f063d17a1
--
NEURODAMUS-CORE Repo SHA: hpc/sim/neurodamus-core@cbb5c827246ef9a3f8f27b4a9797e08a99738893