Further clean up of gap junction branch prior to merging into master
Remove the 'bgq workaround' which is no longer needed.  It turns out not to have been a bgq issue, but rather some nodes just didn't participate in the communication.  Since that is fixed, this code can be safely removed.

Disable the code that was altering the cxinfo file.  It had a bug, but also it is probably not a good idea to change this file format until there is a form of versioning / version checking added.

Insert isc code for reports which did not appear after the merge for some reason.

Restore the activation of cvode.cache_efficient.  Now that gap junctions are debugged, this should work as expected.

JGK
--
NEURODAMUS-CORE Repo SHA: hpc/sim/neurodamus-core@fe17d5bd991856aeea5f15e9a41b916d5f0d7af9