synreader to work when the records have different sizes
SynapseReader mod files read a number of fields depending on the
version and its availability. In case there is a mismatch we now
attempt at coping with it. Is the available fieldset is smaller
then just assign those available. The other way around only happens
when neurodamus-py is outdated, so we raise a warning.

Change-Id: Ib3329fb4964987fb01af6af594e50a1294daf11e