Fix for nrn when sgids are not ascending (!)
In some scenarios, apparently nrn, sgids are not strictly ascending.
This rendered searchsorted inapropriate and among the symptoms was
the incorrect grouping of connections and assignment of properties.
This was noticable in the bonus-minis test where synapses would get
erroneous spont mini rates.

BLUECONFIGS_BRANCH=sandbox/leite/results_update_newparams

Change-Id: I9bd7fab3c6b571b9e9f17c4c0dba5465a236bebf