Fix skipping synapse creation when weight is 0
This patch fixes an unhandled exception raised when there was
a single connection with weight=0, which should simply skip
connection creation altogether.
With this patch the exception doesnt exist anymore and we generalize
the concept by respecting Connection src and dst targets, so that in
case a source of connections (circuit and projections) match a
a single connection block, then the rule applies.

Fixes BBPBGLIB-673

Change-Id: I78e3fb5e82ce21a82e0223a4e8dc278b1e0dfc18