Support for Multi-Circuit
This large refactoring is towards full SONATA multiple
Node and Edge populations. Every "Circuit" block
provides a set of nodes to be simulated, and Projection
blocks specify the source and destination nodes.

To manager the association of edge managers and node managers a new
CircuitManager class was introduced. With this refactoring we
finaly enforce that each node set is handled by an independent
instance of CellManager (CellDistributor for the original METypes),
but further populations can be instantiated. Then each set of
connections between the same populations originate a new
ConnectionManager instante (SynapseRuleManager for the METype
connections, or GapJunctionManager)

As a part of this refatoring comes a number of improvements
in the Engine framework, developed together with NGV.
As so, we see METype as an Engine (the default one) providing
two kinds of connections: SynapseRuleManager and GapJunctions,
similarly to NGV providing NeuroGlial connection managers.

We try really hard to use the real node population names, reading
it if necessary from the Sonata file. However edge files may
reference populations by the Circuit name, and therefore all names
are subject to unaliasing before querying the structures.

EXTRAS
 - Among the cool side effects is that code for specifically hadling GapJunctions
 disappeared, since we generally need to instantiate the class according
 to the connection type, and thus made possible a number of GapJunction
 managers, even possible between different node populations.

 - Dropped handling of legacy option SynapseMode

 - Removed deprecated behavior of searching nrn files inside a dir.
 Dirs are now taken as is, which means either there are NRN-format
 files or it's an error.

Change-Id: Idaa2fbf29b58a5b83d37daa298b679c8ae5d7ed1