Alter check on local_nodes in case of 0 count
For https://bbpteam.epfl.ch/project/issues/browse/BBPBGLIB-745,
when an mpi rank had 0 cells from a population, the local_nodes would report
an offset of 0 rather than the correct count of previous populations.

When preparing reference to local_nodes, we now check the existance of an object
so that it can be used even if there are 0 local cells.

JGK

Change-Id: Ibe9435c42e3338c50a6f96193d3ecec62ec84b7e