Add a GapJunction unit test
This tests bring a minimal sensible simulation with Gap-Junctions and showcases how it could be tested.

We verify that the connection counts are correct, as well as analyzing the voltage values in the soma.
In this test we check the number of peaks of voltage increase rate, but overall we hope to demonstrate the testability of the simulations.