BBPBGLIB-934 LoadBalance refactoring for multiple circuits
## Context

This PR reengineers LoadBalance so that we can keep load balance information on multiple circuits simultaneously, a requirement for BBPBGLIB-934

## Scope

We could get rid of the previous cumbersome logic of having the cxinfo file containing the circuit and the targets. It required a ton of checks to see if things were in sync.
With this refactoring circuits load balance files are permanent, each set staying in their own folder. As such, when checking for availability of load balance one:
 - goes to the folder
 - inspects with cx files are available

And that's it. Folder names are a hash of the nodes file, so there are no clashes (unless the nodes file changes!! which must not happen for production circuits)

## Testing

A new `test_loadbalance.py` file implements a number of unit tests, covering the cases existing in blueconfigs:
 - two synthetic tests
 - Two e2e-style tests, initializing a CellDistributor with real cells and verify load balance properties, inc multispli.
 - And we reached 70% coverage! :tada: 

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or files) are good
* [x] Unit/Scientific test added
* [x] Updated Readme, in-code, developer documentation
* [x] Addressed review comments