BBPBGLIB-954 Avoid flattening nodesets
## Context

To stay compatible with the original targets, NodeSet's were conceived to internally hold an array of the gids. However, `libsonata.Selection` objects are primarily made out of ranges, and flattening million-long ranges (of uint64) takes time and a lot of memory. 

This PR implements a core improvement, and makes neurodamus deal with ranges instead of always flattening the whole array.

This brings important memory and processing savings.

## Scope

This MR introduces a new class of NodeSet's: `SelectionNodeSet`. This is basically a shim for a `libsonata.Selection` object in order to provide the full NodeSet API, namely **Intersection**. Besides the memory savings, we are able to compute intersection more efficiently, using the following new routines:

 - `_ranges_overlap`: Computes the overlap between two lists of ranges. Taking advantage of the fact that ranges come in sorted, the intersection is able to be computed in linear time.
 - `_ranges_vec_overlap`: Detect overlaps between a list of ranges and a vector of ints, eventually the list of local gids which are always a flat list, also in linear time.

## Testing

 - Added two unit tests in `test/test_nodeset.py` to test the new intersection routines.

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or files) are good
* [x] Unit/Scientific test added 
* [ ] Updated Readme, in-code, developer documentation


Closes BBPBGLIB-954