init vascouplingB
## Context

Dan wants to set R0pas based on `vasculature.h5`. It is in the context of [BBPP40-275]

## Scope

Added `_vasculature` to `GlioVascularManager`. It contains a node population. Every node represents a segment of the vasculature. After we populate R0pas in the endfeet segments of the astrocytes

## Testing:

Added tests for Rad and R0pas

## Review
* [ ] PR description is complete
* [ ] Coding style (imports, function length, New functions, classes or files) are good
* [ ] Unit/Scientific test added 
* [ ] Updated Readme, in-code, developer documentation