Improve _iter_conn_params complexity
## Context
Some functions are highly complex and hard to understand/improve.

With the help of Flake we can obtain the McCabe Complexity index, which is an initial indicator. 4 functiona were identified as too complex:
 - ./neurodamus/connection_manager.py:600:5: C901 'ConnectionManagerBase._iterate_conn_params' is too complex (16)
 - ./neurodamus/io/synapse_reader.py:242:5: C901 'SonataReader.preload_data' is too complex (14)
 - ./neurodamus/node.py:762:5: C901 'Node.enable_reports' is too complex (44)
 - ./tests/test_synapses.py:119:1: C901 'test_synapses_params' is too complex (15)

## Scope
This is the first patch to reduce complexity from a central function: `_iterate_conn_params`

## Testing
No new functionality. `_iter_conn_params` is a core routine, exercised in basically every test.

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or files) are good