BBPBGLIB-908 API to restrict features and populations
## Context
In order to allow scientific test we provide an extensible API to selectively enable/disable features and populations.

## Scope
Basically a number of new options can be passed to the top-level `Neurodamus` object.
SimConfig.cli_options understands them and the workflow of Neurodamus respects them.

The options, and respective defaults, are:
  -  restrict_features = NoRestriction  # can also be a list
  -  restrict_node_populations = NoRestriction
  -  restrict_connectivity = 0          # 1 to disable projections, 2 to disable all
  -  restrict_stimulus = NoRestriction  # possible list of Stim names

They can be used like:
```python
    nd = Neurodamus(
        config_file.name,
        restrict_node_populations=["NodeA"],
        restrict_features=[Feature.Replay],
        restrict_connectivity=1,  # base restriction, no projections
        disable_reports=True,
        cleanup_atexit=False,
    )
```


## Testing

A new test is added (test_replay), making use of this feature (The previous example)

## Review
* [ ] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or files) are good
* [x] Unit/Scientific test added 
* [ ] Updated Readme, in-code, developer documentation