Force dtype during numpy.concatenate recarrays
## Context
Due to an incident of the `py-numpy` deployment (BSD-341), we had an issue that `numpy.concatenate` two `numpy.recarray` returned a `numpy.ndarray` of `numpy.void`. It led to an error during the synapse creation because `numpy.void` doesn't maintain type fields (e.g. syn_params.DTC). Forcing `dtype` in `numpy.concatenate` could return `numpy.record` correctly instead of `numpy.void`.

Now with numpy 1.22.4 the issue no more exists. But it is better to merge this MR to protect for the future numpy version.

## Scope
Force to maintain `dtype` in `numpy.concatenate` in `connection.py`

## Testing
Auto fix the failed tests with `v5_sonata` and `usecase3` circuits

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or files) are good
* [ ] Unit/Scientific test added 
* [ ] Updated Readme, in-code, developer documentation