Add test for projections connectivity
## Context
A scientific test for projection connectivity establishment.

## Scope
Added a new test that verifies that projection connectivity is well established and works with replay

It was also noticed that we need absolutely to use the "source" field of the replay stim, otherwise it defaults to the instantiated cell populations (i.e. the _ALL_ target)

## Test
 - Projection Synapses are set VerbosityLevel=1 (SynConfigure). This config only happens in this pathway.
 - Replay is correctly set up and the events reach the target cell (gid 1, pop A)

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or files) are good
* [x] Unit/Scientific test added 
* Updated Readme, in-code, developer documentation