BBPBGLIB-1027 Enable load balance for Sonata config simulations
## Context
As per BBPBGLIB-1027, LoadBalance was not being respected when one uses a Sonata Config.

## Scope
The problem arises since a SONATA config does not define a "Default Circuit".
The fix is to consider the first circuit as the default for Load Balancing.

### Note
This fix meets existing logic (as done for BlueConfigs) where the base circuit is the reference for load balancing.
A much better approach would be to consider all circuits for load balancing, for which a base was already laid out. However, besides more complex, Load balancing as we know might change significantly as per BBPBGLIB-556.

## Testing
Tested locally according to instruction in BBPBGLIB-1027

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or files) are good

Closes BBPBGLIB-1027