Create a testing workflow to run on GitHub actions, to be mirrored to a GitHub branch for SSCx paper
## Context
This MR creates a workflow to be run on GitHub actions which builds the simulation tool chain from scratch and runs a simple simulation test using the local circuit. 

## Scope
The GitHub workflow test contains:
* Install hdf5-mpich
* Install libsonata from source
* Install libsonatareport from source
* Install NEURON
* build h5py with the local hdf5
* Install neurodamus
* Build models from the mod/hoc files provided
* Example Run

It is referenced in `README` as an example.

## Testing
Run test on the local usecase3 circuit, no check on the results.