BBPBGLIB-984 Handle keeping Cell axon during init
## Context
For Multiscale run we need to access properties of the axon sections. However BBP cells usually transform the axon into a 2-section stub thing at load time.

This MR provides the means to stop the cells from replacing the axon, in which cases it will do so before running the simulation.

## Scope
 - Advanced API options can be passed directly to the `Neurodamus` constructor. They can also be set directly in the BlueConfig, per circuit
 - In case the user sets `detailed_axon=True` the setting is passed on to V6 cells init() 3rd argument with the flag value 400.
 - `sim_init()` will, in case, call `.delete_axon()`, in turn calling `replace_axon` on V6 cells.

## Testing
* WIP in a dev scenario under `~leite/test/ngv_wip`

## Review
* [x] PR description is complete
* [ ] Coding style (imports, function length, New functions, classes or files) are good
* [ ] Unit/Scientific test added 
* [ ] Updated Readme, in-code, developer documentation