Replace sys.exit with h.quit to fix issue with MPI_Finalize
## Context
With `sys.exit` the application was forced to exit and didn't call `MPI_Finalize` in the end which especially with `singularity` generated problems.
Now with `h.quit()` `MPI_Finalize` is called and it also returns the proper exit value to the shell.

## Scope
Neuron exit codes and MPI Finalize

## Testing
Should be tests by the CI.

## Review
* [x] PR description is complete
* [ ] Coding style (imports, function length, New functions, classes or files) are good
* [ ] Unit/Scientific test added 
* [x] Updated Readme, in-code, developer documentation