[BBPBGLIB-1036] Implement SONATA synapse reader for gap junction
## Context
Recently we have dropped the synapse-tool dependency for reading synapse parameters. SynReaderNRN is used to read the old format nrn edges files, and SonataReader is to read sonata edges. This PR is to implement a sonata reader to read gap junction synapse parameters.

## Scope
In `gap_junection.py`, a GapJunctionSonata reader derived from the SonataReader. The Sonata parameters for gap junction are `efferent_junction_id` and `afferent_junction_id`.

## Testing
Add a unit test to test GapJunctionSynapseReader for both NRN and SONATA edges files.

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or files) are good
* [x] Unit/Scientific test added 
* [ ] Updated Readme, in-code, developer documentation