import h5py in the function call to filter numpy warnings (#37)
## Context
"special" binaries which flushes denormal results to zero triggers the
numpy warning for subnormal in every rank
after "import h5py". Neurodamus had a filter to ignore this type of
warning, but it is insufficient after the recent change of adding
"import h5py" at the top of replay.py. We have flooded warnings at the
beginning of the simulation before instantiating neuron. This PR changes
to import h5py in the function call to make the filter work again. It
also ensures that we are loading the right h5py version.
 
## Scope
In replay.py, import h5py in the function call instead of at the top.

## Testing
Existing tests

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or
files) are good
* [ ] Unit/Scientific test added
* [ ] Updated Readme, in-code, developer documentation