Print statistics of nodewise memory usage. (#29)
Memory usage statistic of MPI tasks is useful when balancing work on a
per task level. However, for estimating how close one is to going OOM,
the min, max and average don't provide precise information. Instead it
would be useful to see statistics of the memory usage of each compute
node.