Modernize ALU (#48)
**Context**: 

We'd like to make `non_owning_identifier` a non-trivial
class. Because this will allow us to make it reference counted. The
implication is that it'll be important that it's initialized and
destroyed properly.

**Problem**:

ALU.mod contain C code which `malloc`s `data_handle`s, but
`data_handle`s contain a `non_owning_identifier`. Hence, `data_handle`s
must be initialized and destroyed properly.

**Twist**: 

When install NEURON via `pip install` then `nrnivmodl` will
generate a C file `ALU.c` which means we can't use `new` and `delete[]`.

**Fix**: 

Closer inspection of ALU.mod shows that it already requires a
C++ compiler when `NRN_MECHANISM_DATA_IS_SOA` is defined. Since this
covers the cases in which we need a C++ compiler, we use this macro to
conditionally either use `malloc`/`free` or `new`/`delete`.