[HPCTM-1824] Introducing a temporary set for faster lookup in the src_target gids (#63)
## Context
Addresses HPCTM-1824 and HPCTM-1770.

During the synapse configure, a numpy array was created with the gids of
the src_target nodeset for every connection. Then a numpy.searchsorted
was checking if a certain gid was in that array. This was taking a long
time (~1h) when having a lot of connections (~2e8).
The idea is to create a set with the target gids in order to improve the
lookup time.

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or
files) are good
* [ ] Unit/Scientific test added
* [ ] Updated Readme, in-code, developer documentation