[BBPBGLIB-1069] Added nodes suggestions and more improvements (#64)
## Context
Introduces nodes suggestions for users. 
This functionality is one of the main results that should come out of
the introducing of the dry run workflow.

Nodes suggestion is calculated taking in mind the amount of memory
present in the current machine where the code has been run. It also
takes into consideration the variable amount of overhead necessary
dependent on the amount of ranks used for the execution.

## Scope
Most of the changes are independent of the main workflow and only touch
the dry-run workflow. A prompt has been added at the end of any dry run
workflow to provide a suggestion for user on how many nodes are needed
in order to have a reasonable amount of memory available to run the
simulation.

## Testing
Since the MR just introduces a small text prompt for the user, no tests
are necessary.

Full dry run workflow tests will be introduced once the workflow gains a
certain "maturity".

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or
files) are good
* [x] Unit/Scientific test added
* [x] Updated Readme, in-code, developer documentation