[BBPBGLIB-1097] Reorganization of unit/integration tests (#83)
## Context
This PR addresses a long-standing issue of internal tests all being
mixed (apart from scientific ones).

It meant that tests which required more dependencies needed a special
annotation so that they wouldn't run in the CI job for unit tests. It
also meant that the list of loaded modules for these system tests was
hardcoded, and in particular the NGV test had a pretty elaborated hack
around that.

## Scope
We finally segregate the tests according to their nature and
dependencies, and created the main test dirs:

 - tests/unit
 - tests/integration
 - tests/integration-e2e
 - tests/scientific
 - tests/scientific-ngv

With more groups it is also possible to trivially run tests in parallel,
and therefore we dropped pytest-xdist.

Unit tests notably have very little dependencies, even creating and
exposing a NEURON mock for the individual tests.

Also note that `integration-e2e` contain tests that actually run
end-to-end simulations and may rely on blue brain models. However it's
strongly encouraged future integration tests to only rely on NEURON,
which can be placed in `integration` tests.

This PR
 - Moved the tests to their respective group
 - Adapted tests so they don't have any special annotations
 - Adapted testing system and CI, using three TOX groups:
   - unit
   - integration
   - bbp-model
 - Updated contribution docs

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or
files) are good
* [x] Updated Readme, in-code, developer documentation