Remove dependency on mvdtool (#112)
## Context
The last commit has blocked the simulations with BlueConfig files, so we
are now only support SONATA data format. `mvdtool` is used for reading
old mvd3 node files which is no longer needed and can be removed.

## Scope
This PR removes the dependency on `mvdtool` from the package
configuration file (`pyproject.toml`). The relevant source functions
will be cleaned up in the next PRs for further deprecating
legacy/BlueConfig.

A small fix on the unit test regarding the default value of "output_dir" from libsonata 0.1.25

## Testing
Existing tests

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or
files) are good
* [x] Unit/Scientific test added
* [x] Updated Readme, in-code, developer documentation