[BBPBGLIB-1137] Account for nonzero total current when using SEClamp; allow nonzero total current when simulating physical electrodes (#134)
## Context
NEURON reports SEClamp currents at different time points than all other
currents, thus giving inaccurate results in lfp calculations, in cases
where the SEClamp represents synaptic inputs.

This can be fixed by replacing the SEClamp with a new point process
mechanism that is identical, except for producing a NONSPECIFIC_CURRENT
instead of an ELECTRODE_CURRENT.

However, in cases where SEClamp or IClamp are used to represent physical
electrodes, rather than synaptic inputs, the currents should not be
included in the LFP calcualtion

## Scope
When a current_source or conductance_source is used, we check if the
alternative mod file, which is expected to be named
'MembraneCurrentSource.mod' or 'ConductanceSource.mod' is available. If
so, it is used, with nothing else changed, unless the key-value pair
"represents_physical_electrode" is set to True in the Stimulus Config
file. Otherwise the IClamp or SEClamp is used as before.

## Testing
Scientific test added to ensure that new mechanisms produce identical membrane voltages to the old ones

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or
files) are good
* [x] Unit/Scientific test added
* [x] Updated Readme, in-code, developer documentation

---------

Co-authored-by: Jorge Blanco Alonso <jorge.blancoalonso@epfl.ch>
Co-authored-by: Jorge Blanco Alonso <41900536+jorblancoa@users.noreply.github.com>